Create makefile for use in this project
This makefile will compile all our .c files from a directory called src
and places the objects into a directory called out.

To create the executable, make helps to link all the objects files with
gcc and places it inside out as well.

This makefile also allows removing all files in the out folder, allowing
us to recompile everything from scratch if we ever have to.